# Detect logrus exit point when inspecting bugsnag stack