Benchmark without stack inspection:
```
BenchmarkHook_Fire/Error-8         	   50000	     27030 ns/op
BenchmarkHook_Fire/Panic-8         	   50000	     29087 ns/op
```

Benchmark with stack inspection (this patch):
```
BenchmarkHook_Fire/Error-8         	   50000	     29589 ns/op
BenchmarkHook_Fire/Panic-8         	   50000	     31179 ns/op
```

9% slower, seems acceptable for reliability.